make apt-faster support 'apt' in addition to 'apt-get' (#606)
* apt-faster - replace apt as well as apt-get

There are some changes to the wrapper script. The big one
really is just to identify 'full-upgrade' as a "install" operation.

* shellcheck - appease reviewdog

I'm not sure if it will allow the shellcheck disable or not.
It gets confused on 'is never invoked' because those functions
are invoked via the callers of 't()'.