Update setup-go action to v6.0.0
I'm hoping to fix an issue we are seeing privately.
The workflow we are using uses actions/setup-go itself and then
calls setup-melange which also calls setup-go.
The second call (from setup-melange) spews out ~45,000 lines of
errors about `Cannot open: File exists`.

The symptom described is similar to that seen in
https://github.com/actions/setup-go/issues/424

Which is reported to be fixed in v6.0.0.  So, fingers crossed.

Hopefully fixes https://github.com/chainguard-dev/actions/issues/614.